+++ front/.editorconfig
