+++ front/.jshintrc
