+++ front/app/favicon.ico
