+++ front/app/scripts/app.js
