+++ front/app/scripts/controllers/Navbar.js
