+++ front/app/scripts/controllers/ServerDetail.js
