+++ front/app/scripts/controllers/main.js
