+++ front/app/scripts/services/socket.js
