+++ front/app/views/modal/delete.html
