+++ front/app/views/serverdetail.html
