+++ front/app/views/serverlist.html
