+++ front/bower.json
