+++ front/karma-e2e.conf.js
