+++ front/karma.conf.js
