+++ front/test/.jshintrc
