+++ front/test/spec/controllers/Navbar.js
