+++ front/test/spec/controllers/ServerDetail.js
