+++ front/test/spec/controllers/main.js
