+++ front/test/spec/services/socket.js
