--- routes/user.js
@@ -1,8 +1,0 @@
-
-/*
- * GET users listing.
- */
-
-exports.list = function(req, res){
-  res.send("respond with a resource");
-};
